\name{causalBatch-package}
\alias{causalBatch-package}
\alias{causalBatch}
\docType{package}
\title{
\packageTitle{causalBatch}
}
\description{
\packageDescription{causalBatch}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{causalBatch}
\packageIndices{causalBatch}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{causalBatch}

Maintainer: \packageMaintainer{causalBatch}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
