% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_cate.R
\name{cb.sims.cate.heteroskedastic_sim}
\alias{cb.sims.cate.heteroskedastic_sim}
\title{Simulate Data with Heteroskedastic Conditional Average Treatment Effects}
\usage{
cb.sims.cate.heteroskedastic_sim(
  n = 100,
  p = 10,
  pi = 0.5,
  balance = 1,
  eff_sz = 1,
  covar_eff_sz = 5,
  alpha = 2,
  beta = 8,
  common = 10,
  a = 2,
  b = 8,
  err = 1,
  nbreaks = 200
)
}
\arguments{
\item{n}{number of samples to generate}

\item{p}{number of dimensions for the response variable}

\item{pi}{probability of assignment to treatment group 1}

\item{balance}{a parameter that governs the similarity between the covariate distributions}

\item{eff_sz}{effect size parameter controlling the heteroskedasticity between groups}

\item{covar_eff_sz}{effect size parameter for the covariate influence}

\item{alpha}{the alpha parameter for beta distribution of control group}

\item{beta}{the beta parameter for beta distribution of control group}

\item{common}{parameter governing the shape of the common sampling distribution}

\item{a}{scaling factor for response variable generation}

\item{b}{scaling factor for sigmoid transformation}

\item{err}{standard deviation for the error term}

\item{nbreaks}{number of points to use for generating the true signal}
}
\value{
A list containing:
  \item{Ys}{response matrix of size [n, p]}
  \item{Ts}{treatment assignment vector of size [n]}
  \item{Xs}{covariate vector of size [n]}
  \item{Eps}{error matrix of size [n, p]}
  \item{Ytrue}{true response matrix for evaluation}
  \item{Ttrue}{true treatment vector for evaluation}
  \item{Xtrue}{true covariate vector for evaluation}
  \item{Group.Effect}{the effect size parameter}
  \item{Covar.Effect}{the covariate effect size parameter}
}
\description{
Simulate Data with Heteroskedastic Conditional Average Treatment Effects
}
\examples{

library(causalBatch)
sim = cb.sims.cate.heteroskedastic_sim()

}
\references{
Eric W. Bridgeford, et al. "Learning Sources of Variability from High-Dimensional Observational Studies" arXiv (2025).
}
\author{
Eric W. Bridgeford
}
