% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_cate.R
\name{cb.sims.cate.nonmonotone_sim}
\alias{cb.sims.cate.nonmonotone_sim}
\title{Non-monotone CATE Simulation}
\usage{
cb.sims.cate.nonmonotone_sim(
  n = 100,
  p = 10,
  pi = 0.5,
  balance = 1,
  eff_sz = 1,
  alpha = 2,
  beta = 8,
  common = 10,
  err = 1,
  nbreaks = 200,
  rotate = TRUE
)
}
\arguments{
\item{n}{the number of samples. Defaults to \code{100}.}

\item{p}{the number of dimensions. Defaults to \code{10}.}

\item{pi}{the fraction of points which are sampled from a common distribution. Should be a number between \code{0} and \code{1}.}

\item{balance}{a parameter governing the covariate similarity between the two groups. Defaults to \code{1}, which is the same covariate distributions for both groups.}

\item{eff_sz}{the conditional treatment effect size between the different groups, which governs the rotation in radians between the first and second group. Defaults to \code{1}.}

\item{alpha}{the alpha for sampling the \code{0} group, and the beta for sampling the other \code{K - 1} groups.}

\item{beta}{the beta for sampling the \code{0} group, and the alpha for sampling the other \code{K - 1} groups.}

\item{common}{a parameter which governs the shape of the common sampling distribution.}

\item{err}{the level of noise for the simulation. Defaults to \code{1}.}

\item{nbreaks}{the number of breakpoints for computing the expected outcome at a given covariate level
for each batch. Defaults to \code{200}.}

\item{rotate}{whether to apply a random rotation to the outcomes. Defaults to \code{TRUE}.}

\item{covar_eff_sz}{A parameter which governs the covariate effect size with respect to the outcome. Defaults to \code{5}..}
}
\value{
a list, containing the following:
\item{Y}{an \code{[n, p]} matrix, containing the outcomes for each sample.}
\item{Ts}{an \code{[n, 1]} matrix, containing the group labels for each sample.}
\item{Xs}{an \code{[n, 1]} matrix, containing the covariate values for each sample.}
\item{Eps}{an \code{[n, 1]} matrix, containing the error for each sample.}
\item{Ytrue}{an \code{[nbreaks*2, 2]} matrix, containing the expected outcomes at a covariate level indicated by \code{Xtrue}.}
\item{Ttrue}{an \code{[nbreaks*2,1]} matrix, indicating the group/batch the expected outcomes and covariate breakpoints correspond to.}
\item{Xtrue}{an \code{[nbreaks*2, 1]} matrix, indicating the values of the covariate breakpoints for the theoretical expected outcome in \code{Ytrue}.}
\item{Group.Effect}{The group effect magnitude.}
\item{R}{Optional argument returned if a rotation is requested for the rotation matrix applied.}
}
\description{
Non-monotone CATE Simulation
}
\examples{

library(causalBatch)
sim = cb.sims.cate.nonmonotone_sim()

}
\references{
Eric W. Bridgeford, et al. "Learning Sources of Variability from High-Dimensional Observational Studies" arXiv (2025).
}
\author{
Eric W. Bridgeford
}
