% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{cb.sims.covar_generator}
\alias{cb.sims.covar_generator}
\title{Covariate generator function}
\usage{
cb.sims.covar_generator(batches, a1, b1, a2, b2)
}
\arguments{
\item{batches}{an \code{n} vector, consisting of the batch labels for each of the \code{n} samples.}

\item{a1}{alpha of the first covariate distribution.}

\item{b1}{beta of the first covariate distribution.}

\item{a2}{alpha of the second covariate distribution.}

\item{b2}{beta of the second covariate distribution.}
}
\value{
an \code{n} vector, consisting of the covariate values for each of the \code{n} samples.
}
\description{
Covariate generator function
}
