% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{cb.sims.get_beta_overlap}
\alias{cb.sims.get_beta_overlap}
\title{Compute overlap of two beta distributions}
\usage{
cb.sims.get_beta_overlap(a1, b1, a2, b2, nbreaks = 1000)
}
\arguments{
\item{a1}{alpha of the first covariate distribution.}

\item{b1}{beta of the first covariate distribution.}

\item{a2}{alpha of the second covariate distribution.}

\item{b2}{beta of the second covariate distribution.}

\item{nbreaks}{the number of breakpoints for approximating the covariate overlap.}
}
\value{
the level of covariate overlap, corresponding to the AUC upper-bounded
by the probability density functions for each of the beta distributions.
}
\description{
Compute overlap of two beta distributions
}
