% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_ccombat.R
\name{covariate.match}
\alias{covariate.match}
\title{Pairwise covariate matching}
\usage{
covariate.match(covar.tx, covar.cont, match.form, match.args = NULL)
}
\arguments{
\item{covar.tx}{the treatment covariate/label matrix.}

\item{covar.cont}{the control covariate/label matrix.}

\item{match.form}{A formula of columns from \code{Xs}, to be passed directly to \code{\link[MatchIt]{matchit}} for subsequent matching. See \code{formula} argument from \code{\link[MatchIt]{matchit}} for details.}

\item{match.args}{A named list arguments for the \code{\link[MatchIt]{matchit}} function, to be used to specify specific matching strategies, where the list names are arguments and the corresponding values the value to be passed to \code{matchit}.}
}
\value{
a list containing:
\itemize{
   \item{\code{I.mat.k}} index matrix of control samples that have a match.
   \item{\code{M.mat.k}} match matrix of treatment samples that are correspondingly matched to a control.
}
}
\description{
Pairwise covariate matching
}
