% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cdcorr.R
\name{ohe}
\alias{ohe}
\title{A utility to one-hot encode a treatment vector.}
\usage{
ohe(Ts)
}
\arguments{
\item{Ts}{\code{[n]} the labels of the samples, with \code{K < n} levels, as a factor variable.}
}
\value{
\code{[n, K]} a one-hot encoding of \code{Ts}.
}
\description{
A utility to one-hot encode a treatment vector.
}
\author{
Eric W. Bridgeford
}
