% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cdcorr.R
\name{zero_one_dist}
\alias{zero_one_dist}
\title{A utility to compute the zero-one distances for a treatment vector.}
\usage{
zero_one_dist(Ts)
}
\arguments{
\item{Ts}{\code{[n]} the labels of the samples, with \code{K < n} levels, as a factor variable.}
}
\value{
\code{[n, n]} the pairwise zero-one distance matrix.
}
\description{
A utility to compute the zero-one distances for a treatment vector.
}
\author{
Eric W. Bridgeford
}
